+++ code/parachain/frame/assets-registry/src/prelude.rs
@@ -1,0 +1,11 @@
+pub use composable_traits::{
+    assets::Asset,
+    currency::{
+        AssetExistentialDepositInspect, BalanceLike, Exponent,AssetRatioInspect,
+    },
+    defi::Ratio,
+    xcm::assets::{
+        ForeignMetadata, RemoteAssetRegistryInspect,
+        RemoteAssetRegistryMutate,
+    },
+};
